move nav files
