+++ src/components/nav/TheNav.vue
@@ -91,8 +91,8 @@
 
 <script>
 import { mapGetters, mapState } from "vuex";
-import NavLink from "./pure/NavLink.vue";
-import NavMobileMenu from "./pure/MobileMenu.vue";
+import NavLink from "./BaseNavLink.vue";
+import NavMobileMenu from "./MobileMenu.vue";
 import { MenuIcon, XIcon } from "@heroicons/vue/outline";
 
 export default {
